requirements.txt
# AnyArchie - Multi-tenant Personal Assistant
# Python 3.10+

# Telegram bot
python-telegram-bot==21.5

# Database
psycopg[binary]==3.2.4

# LLM
openai==1.58.1

# Web search
exa-py==1.0.9

# PDF generation
reportlab==4.2.5

# Utilities
python-dotenv==1.0.0
httpx==0.28.1
tenacity==9.0.0

# Scheduling
apscheduler==3.10.4

# Google Calendar
google-api-python-client==2.111.0
google-auth==2.27.0

# Encryption
cryptography==42.0.0

# Configuration
pyyaml==6.0.2
